+++ common-db/src/main/java/cn/featherfly/common/db/builder/model/ConditionColumnElement.java
@@ -130,7 +130,9 @@ public class ConditionColumnElement extends ParamedColumnElement {
         if (ComparisonOperator.ISN == comparisonOperator || ComparisonOperator.INN == comparisonOperator) {
             return Params.NONE;
         } else {
-            if (ignore(param)) {
+            if (param instanceof SqlElement) {
+                return super.getParam();
+            } else if (ignore(param)) {
                 return Params.NONE;
             } else {
                 return param;
@@ -146,101 +148,14 @@ public class ConditionColumnElement extends ParamedColumnElement {
         if (Lang.isEmpty(name)) {
             return "";
         }
-        //        StringBuilder condition = new StringBuilder();
         Object value = param;
         String name = dialect.buildColumnSql(tableAlias, this.name);
-        if (ignore(value)) { // 忽略
+        if (value instanceof SqlElement) {
+            return dialect.getCompareExpression(comparisonOperator, name, (SqlElement) value, matchStrategy);
+        } else if (ignore(value)) { // 忽略
             return "";
         } else {
             return dialect.getCompareExpression(comparisonOperator, name, value, matchStrategy);
-            //            if (ComparisonOperator.IN == comparisonOperator || ComparisonOperator.NI == comparisonOperator) {
-            //                int length = 1;
-            //                if (value != null) {
-            //                    if (value instanceof Collection) {
-            //                        length = ((Collection<?>) value).size();
-            //                    } else if (value.getClass().isArray()) {
-            //                        length = Array.getLength(value);
-            //                    }
-            //                }
-            //                condition.append(name).append(Chars.SPACE).append(toOperator(comparisonOperator)).append(" (");
-            //                for (int i = 0; i < length; i++) {
-            //                    if (i > 0) {
-            //                        condition.append(Chars.COMMA);
-            //                    }
-            //                    condition.append(Chars.QUESTION);
-            //                }
-            //                condition.append(")");
-            //            } else if (ComparisonOperator.BA == comparisonOperator || ComparisonOperator.NBA == comparisonOperator) {
-            //                condition.append(name).append(Chars.SPACE)
-            //                        .append(ComparisonOperator.NBA == comparisonOperator
-            //                                ? dialect.getKeyword(Keywords.NOT) + Chars.SPACE
-            //                                : "")
-            //                        .append(dialect.getKeyword(Keywords.BETWEEN)).append(Chars.SPACE) //
-            //                        .append(Chars.QUESTION).append(Chars.SPACE) //
-            //                        .append(dialect.getKeyword(Keywords.AND)).append(Chars.SPACE) //
-            //                        .append(Chars.QUESTION);
-            //            } else {
-            //                condition.append(name).append(Chars.SPACE);
-            //                if (ComparisonOperator.ISN == comparisonOperator) {
-            //                    if ((Boolean) value) {
-            //                        condition.append(toOperator(comparisonOperator));
-            //                    } else {
-            //                        condition.append(toOperator(ComparisonOperator.INN));
-            //                    }
-            //                } else if (ComparisonOperator.INN == comparisonOperator) {
-            //                    if ((Boolean) value) {
-            //                        condition.append(toOperator(comparisonOperator));
-            //                    } else {
-            //                        condition.append(toOperator(ComparisonOperator.ISN));
-            //                    }
-            //                } else {
-            //                    condition.append(toOperator(comparisonOperator)).append(Chars.SPACE).append(Chars.QUESTION);
-            //                }
-            //            }
         }
-        //        return condition.toString();
     }
-
-    //    private String toOperator(ComparisonOperator comparisonOperator) {
-    //        switch (comparisonOperator) {
-    //            case EQ:
-    //                return dialect.getOperators().eq();
-    //            case NE:
-    //                return dialect.getOperators().ne();
-    //            case SW:
-    //                return dialect.getKeywords().like(matchStrategy);
-    //            case NSW:
-    //                return dialect.getKeywords().notLike(matchStrategy);
-    //            case CO:
-    //                return dialect.getKeywords().like(matchStrategy);
-    //            case NCO:
-    //                return dialect.getKeywords().notLike(matchStrategy);
-    //            case EW:
-    //                return dialect.getKeywords().like(matchStrategy);
-    //            case NEW:
-    //                return dialect.getKeywords().notLike(matchStrategy);
-    //            case LK:
-    //                return dialect.getKeywords().like(matchStrategy);
-    //            case NL:
-    //                return dialect.getKeywords().notLike(matchStrategy);
-    //            case LT:
-    //                return dialect.getOperators().lt();
-    //            case LE:
-    //                return dialect.getOperators().le();
-    //            case GT:
-    //                return dialect.getOperators().gt();
-    //            case GE:
-    //                return dialect.getOperators().ge();
-    //            case IN:
-    //                return dialect.getKeywords().in();
-    //            case NI:
-    //                return dialect.getKeywords().notIn();
-    //            case ISN:
-    //                return dialect.getKeywords().isNull();
-    //            case INN:
-    //                return dialect.getKeywords().isNotNull();
-    //            default:
-    //                return dialect.getOperators().eq();
-    //        }
-    //    }
 }