+++ common-db/src/main/java/cn/featherfly/common/db/dialect/MySQLDialect.java
@@ -12,6 +12,7 @@ import cn.featherfly.common.constant.Chars;
 import cn.featherfly.common.db.Column;
 import cn.featherfly.common.db.SqlUtils;
 import cn.featherfly.common.db.builder.BuilderUtils;
+import cn.featherfly.common.db.builder.model.SqlElement;
 import cn.featherfly.common.lang.ArrayUtils;
 import cn.featherfly.common.lang.AssertIllegalArgument;
 import cn.featherfly.common.lang.Dates;
@@ -36,6 +37,7 @@ public class MySQLDialect extends AbstractDialect {
      * Instantiates a new my SQL
      */
     public MySQLDialect() {
+        super();
     }
 
     /**
@@ -92,11 +94,10 @@ public class MySQLDialect extends AbstractDialect {
         pagingSelect.append(sql);
         if (isParamNamed) {
             if (start > 0) {
-                pagingSelect.append(Strings.format(" LIMIT {0}{1},{0}{2}",
-                        Lang.array(PARAM_NAME_START_SYMBOL, START_PARAM_NAME, LIMIT_PARAM_NAME)));
+                pagingSelect.append(
+                        Strings.format(" LIMIT {0}{1},{0}{2}", startSymbol, START_PARAM_NAME, LIMIT_PARAM_NAME));
             } else {
-                pagingSelect
-                        .append(Strings.format(" LIMIT {0}{1}", Lang.array(PARAM_NAME_START_SYMBOL, LIMIT_PARAM_NAME)));
+                pagingSelect.append(Strings.format(" LIMIT {0}{1}", startSymbol, LIMIT_PARAM_NAME));
             }
         } else {
             if (start > 0) {
@@ -368,6 +369,37 @@ public class MySQLDialect extends AbstractDialect {
      * {@inheritDoc}
      */
     @Override
+    public String getBetweenOrNotBetweenExpression(boolean isBetween, String name, SqlElement min, SqlElement max,
+            MatchStrategy matchStrategy) {
+        boolean caseSensitive = false;
+        switch (matchStrategy) {
+            case CASE_INSENSITIVE:
+                break;
+            case CASE_SENSITIVE:
+                caseSensitive = true;
+                break;
+            default:
+                return getBetweenOrNotBetweenExpression(isBetween, name, min, max);
+        }
+        StringBuilder condition = new StringBuilder();
+        if (caseSensitive) {
+            condition.append(getKeyword(Keywords.BINARY)).append(Chars.SPACE).append(name);
+        } else {
+            condition.append(name).append(Chars.SPACE).append(getKeyword(Keywords.COLLATE)).append(Chars.SPACE)
+                    .append(collateCaseInsensitive);
+        }
+        condition.append(Chars.SPACE).append(!isBetween ? getKeyword(Keywords.NOT) + Chars.SPACE : "") //
+                .append(getKeyword(Keywords.BETWEEN)).append(Chars.SPACE) //
+                .append(min.toSql()).append(Chars.SPACE) //
+                .append(getKeyword(Keywords.AND)).append(Chars.SPACE) //
+                .append(max.toSql());
+        return condition.toString();
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
     protected String getCompareExpression0(ComparisonOperator comparisonOperator, String name, Object values,
             MatchStrategy matchStrategy) {
         switch (comparisonOperator) {
@@ -408,4 +440,24 @@ public class MySQLDialect extends AbstractDialect {
         return condition.toString();
     }
 
+    @Override
+    protected String getCompareExpression0(ComparisonOperator comparisonOperator, String name, SqlElement values,
+            MatchStrategy matchStrategy) {
+        StringBuilder condition = new StringBuilder();
+        switch (matchStrategy) {
+            case CASE_INSENSITIVE:
+                condition.append(name).append(Chars.SPACE).append(getKeyword(Keywords.COLLATE)).append(Chars.SPACE)
+                        .append(collateCaseInsensitive);
+                break;
+            case CASE_SENSITIVE:
+                condition.append(getKeyword("BINARY")).append(Chars.SPACE).append(name);
+                break;
+            default:
+                condition.append(name);
+                break;
+        }
+        condition.append(Chars.SPACE).append(getOperator(comparisonOperator)).append(Chars.SPACE)
+                .append(values.toSql());
+        return condition.toString();
+    }
 }