+++ common-db/src/test/java/cn/featherfly/common/db/dialect/DialectTest.java
@@ -215,6 +215,8 @@ public abstract class DialectTest {
 
     abstract void testBuildInsertBatchSql();
 
+    abstract void testPaginationSql();
+
     abstract void testParamNamedPaginationSql();
 
     abstract void testInsert();